bytes2str: fix slowness when passed binary data
As discovered through performance profiling, bytes2str is what is
causing binary uploads to be slow and ultimately fail with a timeout.

Disabling the log output on each invalid UTF-8 sequence fixes the
slowness.

Instead, the number of replacements is counted and a summary message
is printed at the end.